requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
backports.tarfile==1.2.0
bcrypt==3.2.2
beautifulsoup4==4.13.0
blinker==1.8.1
build==1.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
datasets==4.2.0
decorator==4.4.2
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docutils==0.22.2
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
facenet-pytorch==2.5.3
fastapi==0.119.1
fer==22.5.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.75.1
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
id==1.5.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
intel-openmp==2021.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
keras==3.11.3
keyring==25.6.0
kiwisolver==1.4.9
libclang==18.1.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
networkx==3.4.2
nh3==0.3.1
numpy==2.2.6
openai==2.6.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pefile==2023.2.7
pillow==10.2.0
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
readme_renderer==44.0
regex==2025.10.22
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tbb==2021.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-intel==0.0.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.1
twine==6.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
